Rename torch adapter file to avoid mypy failures
